backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.20
alembic>=1.12.0
asyncpg>=0.29.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic[email]>=2.6.0
stripe>=7.4.0
twilio>=8.10.0
sendgrid>=6.10.0
openpyxl>=3.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
